build: refactored build process
fixes #106, #80, #82
