+++ .babelrc
@@ -1,11 +1,4 @@
 {
-  "presets": [
-    ["env"],
-    "react"
-  ],
-  "plugins": [
-    "transform-object-rest-spread",
-    "transform-class-properties",
-    "transform-export-extensions"
-  ]
+  "presets": [["env", { "modules": false }], "stage-2", "react"],
+  "plugins": ["transform-class-properties"]
 }
