+++ .gitignore
@@ -21,3 +21,4 @@ yarn-error.log
 .project
 .tmp
 .vscode
+*.log
