+++ demo/index.html
@@ -1,12 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta charset="UTF-8">
   <title>Component</title>
   <link href="https://fonts.googleapis.com/css?family=Bungee|Lato:400,900" rel="stylesheet">
 </head>
+
 <body>
   <div id="app"></div>
-  <script src="bundle.js"></script>
+  <script src="./index.js"></script>
 </body>
+
 </html>
