+++ package.json
@@ -1,18 +1,36 @@
 {
   "name": "react-popper",
-  "version": "0.8.3",
+  "version": "0.9.0",
   "description": "React wrapper around PopperJS.",
+  "license": "MIT",
+  "author":
+    "Travis Arnold <travis@souporserious.com> (http://souporserious.com)",
+  "homepage": "https://github.com/souporserious/react-popper",
   "main": "dist/react-popper.js",
+  "umd:main": "dist/react-popper.umd.js",
   "module": "lib/react-popper.js",
   "types": "react-popper.d.ts",
-  "files": [
-    "dist",
-    "lib",
-    "react-popper.d.ts"
-  ],
-  "license": "MIT",
-  "author": "Travis Arnold <travis@souporserious.com> (http://souporserious.com)",
-  "homepage": "https://github.com/souporserious/react-popper",
+  "files": ["dist", "lib"],
+  "scripts": {
+    "build":
+      "npm run build:clean && npm run build:es && npm run build:umd && npm run build:cjs && npm run build:umd-min && npm run build:cjs-min",
+    "build:clean": "rm -rf dist/ && rm -rf lib/",
+    "build:es":
+      "babel src --out-dir lib && mv lib/index.js lib/react-popper.js",
+    "build:umd":
+      "rollup -c --output.format umd --output.name 'react-popper' --output.file dist/react-popper.umd.js",
+    "build:cjs":
+      "rollup -c --output.format cjs --output.name 'react-popper' --output.file dist/react-popper.js",
+    "build:umd-min":
+      "MINIFY=true rollup -c --output.format umd --output.name 'react-popper' --output.file dist/react-popper.umd.min.js",
+    "build:cjs-min":
+      "MINIFY=true rollup -c --output.format cjs --output.name 'react-popper' --output.file dist/react-popper.min.js",
+    "demo": "parcel --out-dir demo/dist demo/index.html",
+    "prepare": "npm run build",
+    "precommit": "lint-staged",
+    "format": "npm run lint -- --fix --quiet",
+    "lint": "eslint --ext .js,.jsx src/ example/"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/souporserious/react-popper"
@@ -30,21 +48,7 @@
     "popover"
   ],
   "lint-staged": {
-    "@(src|example)/**/*.{js,jsx}": [
-      "eslint --fix --max-warnings 5",
-      "git add"
-    ]
-  },
-  "scripts": {
-    "precommit": "lint-staged",
-    "format": "npm run lint -- --fix --quiet",
-    "lint": "eslint --ext .js,.jsx src/ example/",
-    "build:lib": "babel src --out-dir lib",
-    "build": "npm run build:lib && cross-env NODE_ENV=production webpack --config webpack.prod.config.js",
-    "dev": "webpack-dev-server --hot --progress --host 0.0.0.0",
-    "postbuild": "cross-env NODE_ENV=production && cross-env TARGET=minify webpack --config webpack.prod.config.js",
-    "prebuild": "rimraf dist && mkdir dist",
-    "prepare": "npm run build"
+    "@(src|demo)/**/*.{js,jsx}": ["eslint --fix --max-warnings 5", "git add"]
   },
   "peerDependencies": {
     "react": "0.14.x || ^15.0.0 || ^16.0.0",
@@ -57,36 +61,33 @@
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.0",
-    "babel-eslint": "^8.2.2",
-    "babel-loader": "^7.1.4",
+    "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-export-extensions": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.13",
     "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.16.0",
-    "cross-env": "^5.1.4",
-    "css-loader": "^0.28.11",
-    "eslint": "^4.19.0",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-config-standard": "^11.0.0",
-    "eslint-config-standard-react": "^6.0.0",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-prettier": "^2.6.0",
-    "eslint-plugin-promise": "^3.7.0",
-    "eslint-plugin-react": "^7.7.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "husky": "^0.14.3",
-    "lint-staged": "^7.0.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "chokidar": "^1.6.1",
+    "create-styled-element": "^0.4.0",
+    "glamor": "^2.20.24",
+    "glamorous": "^3.11.1",
+    "node-libs-browser": "^1.0.0",
     "outy": "^0.1.2",
-    "prettier": "^1.11.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-portal": "^4.1.3",
+    "parcel-bundler": "^1.6.2",
+    "react": "15.3.2",
+    "react-dom": "15.3.2",
+    "react-measure": "^1.4.6",
+    "react-motion-ui-pack": "^0.10.2",
+    "react-portal": "^4.1.4",
     "react-transition-group": "^2.2.1",
-    "rimraf": "^2.6.2",
-    "style-loader": "^0.20.3",
-    "webpack": "^3.11.0",
-    "webpack-dev-server": "^2.11.2"
+    "react-travel": "^1.3.0",
+    "rollup": "^0.57.1",
+    "rollup-plugin-babel": "^3.0.3",
+    "rollup-plugin-babel-minify": "^4.0.0",
+    "rollup-plugin-commonjs": "^9.1.0",
+    "rollup-plugin-node-globals": "^1.2.0",
+    "rollup-plugin-node-resolve": "^3.3.0",
+    "rollup-plugin-replace": "^2.0.0",
+    "velocity-react": "^1.2.2"
   }
 }
