+++ src/Manager.js
