+++ src/index.js
@@ -1,0 +1,4 @@
+export { default as Manager } from './Manager';
+export { default as Target } from './Target';
+export { default as Popper, placements } from './Popper';
+export { default as Arrow } from './Arrow';
