--- src/react-popper.js
@@ -1,4 +1,0 @@
-export Manager from './Manager'
-export Target from './Target'
-export Popper, { placements } from './Popper'
-export Arrow from './Arrow'
